vim-patch:9f2962141514
Runtime(termdebug): Add support to view local and argument variables

closes: 12403

https://github.com/vim/vim/commit/9f29621415146abc046471440515e9e34f3e57a1

Rename the existing "s:running" (#16790) to "s:gdb_running" to not clash with
the "s:running" introduced in this patch (which instead relates to whether the
debugged program is currently running in gdb).

Keep the file `:retab`bed as before.

Co-authored-by: laburnumT <flo.striker@gmail.com>