fix(runtime): make `:Sman/:{Get,Run}Help` work in Nvim
Problem: vim-patch:2f25e40b1f54 may use `less` interactively via :! in :Sman and
:Help, which doesn't work in Nvim. This made me also notice there's
:{Get,Run}Help in other files...

Solution: Use :terminal. :split by default (like Vim) so it works well with K,
and :startinsert so that spawned pagers can be interacted with immediately.

Maybe :Man is better, but this is simpler and diverges less (also the
"--pattern"s used in `less` would need to be adjusted to Vim syntax).