feat(lsp/buf): optionally suppress no LSP notification
This makes it possible to suppress the warning indicating that no LSP
matching LSP servers were found. This can happen if you have a filter
that would exclude one of the servers. For example you are excluding
tsserver with the intention that you would use eslint instead. But if
eslint is not installed or configured for that project then you would
be able to silence the warning.