fix(tests): ignore $SHELL value in tests
Fixes #6172

"echo -n" was replaced by "printf", because "echo" in sh may be provided
as a shell builtin, which does not accept an "-n" flag to avoid a
trailing newline (e.g. on macos). "printf" is being used instead,
because it's more portable (defined by POSIX) and it does not output a
trailing newline by itself.