refactor(languagetree): deprecate `for_each_child`
The name was misleading and caused bugs. After fixing these bugs
(#25111, #25115), there are no more usages of `for_each_child`.

Since the function is technically part of the external API, I've opted
not to outright remove it yet.

In the future, if we want to restore this functionality, we can add to
`shared.lua` something like:

```lua
--- Visits children of a node recursively, applying the provided visitor function
---@generic Node: table
---@param node Node The root node to visit
---@param visitor fun(node: Node): boolean|nil Function to call at each node.
---       If it returns true, it will skip recursing into the children of the node.
---@param key nil|string|fun(node: Node): Node[] How to find the children of a node.
function vim.traverse(node, visitor, key)
  if not visitor(node) then
    local children = node.children
    if type(key) == 'function' then
      children = key(node)
    elseif key then
      children = node[key]
    end
    for _, child in pairs(children or {}) do
      vim.traverse(child, visitor, key)
    end
  end
end
```