fix(treesitter): add Query:captures_iterator()
07080f67f introduced changes to `iter_captures()` by adding arguments to
the created iterator to control each iteration step. However this breaks
usages of `iter_captures()` in for-loops since the arguments of
iterators should be the invariant state and control variable, see:
https://www.lua.org/pil/7.3.html.

This change fixes that by introducing a new function
`captures_iterator()` which by design cannot be used in for-loops, and
restores `iter_captures()` to return a stateless iterator (function with
no arguments).