fix(languagetree): cache `resolve_lang` to avoid thrashing filesystem
`vim._ts_has_language` works brilliantly as a cache when the parser file
is found. However, if no parser file is found, `has_parser` will keep
trying to open the file, which will block the main thread with slow I/O
operations, leading to significant lag on, e.g. `languagetree.lua`.

I found this out when I uninstalled the `comment` parser (should have
been a performance bump but interestingly revealed this bug).

The cache is scoped to the `LanguageTree` instance, and the only way to
effectively cache bust is to create a new `LanguageTree`. I'm open to a
different cache policy, ideas are welcome.