fix(lsp): fix omnicomplete in middle of the line
I experienced that no results were displayed from lsp server when I
tried to autocomplete from the middle of a line.

I tracked down _completion.lua `M._lsp_to_complete_items` function, where
the matches_prefix didn't match to anything, because the prefix was
always the whole line and not stopped at the cursor.

I found that the prefix `M.convert_results function`:
`local prefix = line:sub((server_start_boundary or client_start_boundary) + 1)`

Probably it's a regression from 5e5f5174e3faa862a9bc353aa7da41487911140b

Until that commit we had a logic like this:
`local prefix = startbyte and line:sub(startbyte + 1) or line_to_cursor:sub(word_boundary)`

So in case of `startbyte` was falsy, we had a different logic and used
the `line_to_cursor` instead if line.

To fix the issue, I pass the `cursor_column` to the
`M._convert_results` function. Then when we construct the prefix we use
it as the last character index. So the prefix will start at the word
boundary and ends at the cursor.

I choose to pass the cursor_column instead of the line, so the
`_convert_results` function holds the logic and it's testable