docs(api): add User autocmd example
It's a common misconception about vimscript that the User autocmd is
special, and many people don't realize that the event name given to the
autocommand is just a regular pattern.

To eliminate this misconception, and to make it clear how to make a User
autocommand, an example is added to illustrate that the event name is
just a regular pattern.