refactor(api)!: redesign `nvim_(get|set)_option_value`
Problem: Currently, `nvim_(get|set)_option_value` can modify the local value forthe current buffer or window even when it's not explicitly specified.This is undesirable behavior for the API, as any changes to a localvalue should only happen when the intent to do so is explicitlyexpressed by the caller of the API function.

Solution: Make `nvim_(get|set)_option_value` always require window or buffer to be explicitly specified when a local value needs to be modified. Also change how default scope is picked to always prefer global scope by default whenever possible. Also add a `'both'` scope to set both the local and global scope.

NOTE: #25674 is a prerequisite for this.