feat(clipboard): enable OSC 52 clipboard provider by default
Use the XTGETTCAP sequence to determine if the host terminal supports
the OSC 52 sequence and, if it does, enable the OSC 52 clipboard
provider by default.

The behavior is controlled by the g:osc52 global variable. If unset/nil,
then it is only enabled by default when all of the following are true:

  1. Nvim is running in an SSH connection ($SSH_TTY is set)
  2. Nvim is not running inside tmux ($TMUX is unset)
  3. 'clipboard' is not set to unnamed or unnamedplus

If g:osc52 is true then the conditions above are ignored and the OSC 52
provider is always enabled by default (if the terminal supports it). If
g:osc52 is false then the provider is never enabled by default.