feat(treesitter): add `#if?` predicate for conditional queries
Problem: There is no way to control queries at runtime, meaning
controversial conceals or expensive injections have to be left to user
config (and are not possible to conditionally enable). This is a
regression over legacy syntax, which gates some features behind
configuration variables.

Solution: Add an `#if?` predicate that can be used to gate patterns
behind a configuration variable, where buffer variables take precedence.
For example, the pattern
```
((identifier) @foo (#if? "identifier-foo"))
```
is only applied if
1. `vim.b.identifier-foo` is truthy or
2. `vim.g.identifier-foo` is truthy and `vim.b.identifier-foo` is not
   falsy.